+++ src/main/java/com/openggf/trace/TraceExecutionPhase.java
@@ -2,5 +2,6 @@ package com.openggf.trace;
 
 public enum TraceExecutionPhase {
     FULL_LEVEL_FRAME,
+    ADVANCE_ONLY,
     VBLANK_ONLY
 }